Restore dewplayer for audio previews.
This also starts the work to have the preview/details html structures generated by the front end where it belongs, not by the PHP backend which should be limited to sending data, not markup.

Note that if you are "restricting downloads" the dewplayer preview obviously may not work depending on how exactly you are doing the restricting.