Scratch that, this is a much better idea, preview handlers can now be specified in the options.
You can specify a full or partial mime type (eg audio, or audio/mpeg) and an associated function which will get the DOM element to stick your preview stuff, and the file information returned from the backend.  By default we add in our own audio handler if none other is supplied.  The other media types should eventually be handled in the same manner.