Add a new MP3 id3 tag extractor class.  Developed for Mootools File Manager and MIT licensed.
The class operates on a "good enough" basis, that is, it does as much as we practically need in Mootools File Manager (although perhaps album art could be added).

It is quite robust, running it over my collection of MP3 files (circa 6000 files) produces no issues.

It handles version 1, 2.2, 2.3 and 2.4 tags.

"Unsynchronised" tags are handled, but the specification for this is a bit murky so it might not be perfect.

Encrypted and/or compressed tags are not handled (they are silently skipped).

Multiple values in a single "frame" (eg multiple artists) are NOT returned as an array (as I think getID3 does), the value of an
id3 tag is presently always going to be a string.  A UTF-8 string indeed.  The FileManager.php has been modified to accomodate
this.