Downloads throw a 500 server error
PHP can throw a 500 server error when the files trying to be downloaded are larger than PHP memory limit. This fix will use PHP's location: header to pass the file download off to the web server. This will negate any PHP timeouts (or 500 errors). This fix will also negate the need for the frontend's iframe downloader. Downloads instead will be handled in a new window.