Elegantly handle hosts with suhosin's cookie encryption.
suhosin encrypts/decrypts the PHP cookies on the fly. This seems to be
transparent to PHP. Flash, however, will propogate the encrypted cookie
in the propagateData object. As a result, if suhosin.cookie.encrypt is
enabled, we can't trust propagateData's value for $sesname. Therefore,
we should use something else. By default, don't expose the session id
if we don't have to.