Various small changes implemented locally over the years.
Remove space ( ) from the list of allowed characters in filenames.  They will be converted to underscore (_) along with other disallowed characters.  Spaces in file names create more trouble than they are worth.

Fix for extention safety with regard to duplicate files.  Previously if an extention was changed by the getSafeExtension (eg JPG to jpg) then uploading a duplicate file with the JPG extention would result in an error or overwrite rather than appending _1.jpg etc, due to an ordering of the code.

Add some more php extensions to the unsafe list.

When images were set to resize on upload if too large the resized image was done at 100 quality and stored regardless if the original met the requirements or not.  This fixes that with quality set to 60 and only stored if it had to be resized.  Production experience has shown this to be generally acceptable, if people need higher qualtity they can upload correctly sized images.

Prevent double ampersands in mkServerRequestURL responses, I forget why but I guess it must have caused me a problem at some point.